+++ templates/admin.php
@@ -144,7 +144,8 @@ function print_select_options(
                 print_select_options($l, "opt-email_sync", "Email sync", ["" => "None", "initial" => "Synchronise only once", "force_nc"=>"Nextcloud always wins", "force_sql"=>"SQL always wins"], $_["opt.email_sync"]);
                 print_select_options($l, "opt-quota_sync", "Quota sync", ["" => "None", "initial" => "Synchronise only once", "force_nc"=>"Nextcloud always wins", "force_sql"=>"SQL always wins"], $_["opt.quota_sync"]);
                 print_select_options($l, "opt-home_mode", "Home mode", ["" => "Default", "query" => "Query", "static" => "Static"], $_["opt.home_mode"]);
-                print_text_input($l, "opt-home_location", "Home Location", $_["opt.home_location"]); ?>
+                print_text_input($l, "opt-home_location", "Home location", $_["opt.home_location"]);
+                print_text_input($l, "opt-default_group", "Default group", $_["opt.default_group"]); ?>
             </fieldset>
         </div>
         <div class="section clear-left">